requirements.txt
streamlit
pillow
numpy
pandas
scikit-learn
opencv-python-headless
torch
torchvision
kaggle
matplotlib
